+++ cudajit.mli
@@ -8,8 +8,8 @@ module Nvrtc : sig
 
   exception Nvrtc_error of { status : result; message : string }
   (** Error codes returned by CUDA functions are converted to exceptions. The message stores a
-      snake-case variant of the offending CUDA function name (see {!Cuda_ffi.C.Functions} and
-      {!Nvrtc_ffi.C.Functions} for the direct funciton bindings). *)
+      snake-case variant of the offending CUDA function name (see {!Nvrtc_ffi.Bindings.Functions}
+      for the direct funciton bindings). *)
 
   type compile_to_ptx_result [@@deriving sexp_of]
   (** The values passed from {!compile_to_ptx} to {!module_load_data_ex}. Currently, cudajit
@@ -49,8 +49,12 @@ type result [@@deriving sexp]
 
 exception Cuda_error of { status : result; message : string }
 (** Error codes returned by CUDA functions are converted to exceptions. The message stores a
-    snake-case variant of the offending CUDA function name (see {!Cuda_ffi.C.Functions} and
-    {!Nvrtc_ffi.C.Functions} for the direct funciton bindings). *)
+    snake-case variant of the offending CUDA function name (see {!Cuda_ffi.Bindings.Functions} for
+    the direct funciton bindings). *)
+
+val cuda_call_callback : (message:string -> status:result -> unit) option ref
+(** The function called after every {!Cuda_ffi.Bindings.Functions} call. [message] is the snake-case
+    variant of the corresponding CUDA function name. *)
 
 val init : ?flags:int -> unit -> unit
 (** Must be called before any other function. Currently [flags] is unused. See