+++ test/saxpy.ml
@@ -20,6 +20,7 @@ let%expect_test "SAXPY" =
     Cu.Nvrtc.compile_to_ptx ~cu_src:kernel ~name:"saxpy" ~options:[ "--use_fast_math" ]
       ~with_debug:true
   in
+  Cu.cuda_call_callback := Some (fun ~message ~status:_ -> Printf.printf "%s\n" message);
   Cu.init ();
   if Cu.Device.get_count () > 0 then (
     let device = Cu.Device.get ~ordinal:0 in
@@ -65,6 +66,25 @@ let%expect_test "SAXPY" =
     Format.print_newline ();
     [%expect
       {|
+      cu_init
+      cu_device_get_count
+      cu_device_get
+      cu_ctx_create
+      cu_module_load_data_ex
+      cu_module_get_function
+      cu_mem_alloc
+      cu_memcpy_H_to_D
+      cu_mem_alloc
+      cu_memcpy_H_to_D
+      cu_mem_alloc
+      cu_memcpy_H_to_D
+      cu_launch_kernel
+      cu_ctx_synchronize
+      cu_memcpy_D_to_H
+      cu_mem_free
+      cu_mem_free
+      cu_mem_free
+      cu_module_unload
       5.1 * 0.0 + 0.0 = 0.00; 5.1 * 1.0 + 2.0 = 7.10; 5.1 * 2.0 + 4.0 = 14.20; 5.1 * 3.0 + 6.0 = 21.30;
       5.1 * 4.0 + 8.0 = 28.40; 5.1 * 5.0 + 10.0 = 35.50; 5.1 * 6.0 + 12.0 = 42.60; 5.1 * 7.0 + 14.0 = 49.70;
       5.1 * 8.0 + 16.0 = 56.80; 5.1 * 9.0 + 18.0 = 63.90; 5.1 * 10.0 + 20.0 = 71.00; 5.1 * 11.0 + 22.0 = 78.10;
@@ -1428,4 +1448,5 @@ let%expect_test "SAXPY" =
       5.1 * 4084.0 + 8168.0 = 28996.40; 5.1 * 4085.0 + 8170.0 = 29003.50; 5.1 * 4086.0 + 8172.0 = 29010.60;
       5.1 * 4087.0 + 8174.0 = 29017.70; 5.1 * 4088.0 + 8176.0 = 29024.80; 5.1 * 4089.0 + 8178.0 = 29031.90;
       5.1 * 4090.0 + 8180.0 = 29039.00; 5.1 * 4091.0 + 8182.0 = 29046.10; 5.1 * 4092.0 + 8184.0 = 29053.20;
-      5.1 * 4093.0 + 8186.0 = 29060.30; 5.1 * 4094.0 + 8188.0 = 29067.40; 5.1 * 4095.0 + 8190.0 = 29074.50; |}])
+      5.1 * 4093.0 + 8186.0 = 29060.30; 5.1 * 4094.0 + 8188.0 = 29067.40; 5.1 * 4095.0 + 8190.0 = 29074.50;
+      |}])